Implement q_remove_head and q_remove_tail
It may need to use the node removal function multiple times in the
future, so I added a function remove_element to handle this job, the
functino will get the element_t object which containing the node is
prepare to be removed, copy the value, delete the node from linked
list then return the result.

In q_remove_head and q_remove_tail, the corresponding positions will be
directly passed into remove_element.