Implement q_delete_mid and fix typo
Implement q_delete_mid, by using 2 pointers with different moving speed.
When the fastest pointer is at the end of the linked list, the other one
will stop at the middle of the list, then delete that node and the
element_t object which contains it.

Fix typos on line 16 and line 29.