Implement q_delete_dup
Implement q_delete_dup, using list_for_each_safe to get every nodes in
the linked list, compares the value of the element_t object containing
the current node with the value of the next node, if they are the same
delete the current one and mark the next node for deleton.
I use a Boolean variable last as a marker, if the marker is true the
current node will be deleted no matter what. The marker will be reset
when the value of the element_t object is different from the next node.