Implement q_ascend and q_descend
Implement q_ascend and q_descend, the linked list will be reversed
before the process and reverses it back to the original order. It will
be easier to process if the linked list is reversed.
When processing q_ascend, the value of the element_t object containing
current node will be compared with the minimum value among the currently
processed nodes. If the currently value is larger, the node will be
deleted, otherwise the data will be recorded.
This also works for q_descend, but records the maximum value instead and
removes smaller nodes.