Implement q_sort, fix bug in q_delete_dup & q_swap
Impement q_sort by using merge sort, the input list will be split into 2
equal parts and sort them (also use q_sort), they will be merged after
both of them are sorted.
The way to slit the input linked list is same as q_delete_mid, the first
part will storaged in left and the rest will be left in head.
I choose to find position in head for each elements in left, if the
pointer of head are alreay at the tail, the elements in left that have
not yet been placed will be put at the end of head.

There is a bug in q_delete_dup: all elementsthat follow the first group
of identical elements will be deleted in linked list.
This is because there is a Boolean variable last will not be changed as
long as it is changed to true (the contition in line 161 never be false
since last has been set to true).
To fix this bug, I choose to sync the value with match result at the end
of this check.

There is also another bug in q_swap, I found it when I was testing this
command with a linked list have 3 elements and k was 2, the last
elements should not be changed but the test result didn't show the
desired effect, so I just replace the feature with a k = 2 q_reverseK.