Implement q_shuffle
Implement q_shuffle with Fisher-Yates shuffle algorithm and a swap
function to swap 2 nodes.

The original process in Fisher-Yates shuffle algorithm would start from
the end of the list, but I decided to start from the head, because it
did not matter.

TODO: The random function needs to work on.